requirements.txt
faster-coco-eval>=1.6.5
PyYAML
tensorboard
scipy
calflops
thop
transformers
pytorch_wavelets==1.3.0
timm==1.0.7
grad-cam==1.5.4
tidecv
einops
prettytable

# ------------------ 装不成功可以先注释这两个
mmengine==0.10.7
mmcv==2.2.0
